Business Analytics Capstone Consultancy business Analytics project
This Microsoft-Word-based report is one of the requirements of the Capstone consultancy project towards completion of my 1-year Business Analytics postgraduate course at Seneca College, Toronto. It showcases not only my ability to handle large and complex data sets and draw actionable insights but also lead business Analytics projects and Teams effectively. It was a great opportunity to apply and showcase all the concept learned during the course and their application to a real-life situation. "Knowledge is not Power until applied"! Dale Carnegie.